database: cutomer repo

+++ src/badrbillingsystem/repos/customer/ICustomerRepo.java
@@ -15,8 +15,8 @@ import java.util.ArrayList;
 public interface ICustomerRepo {
     boolean save(Customer customer);
     boolean update(Customer newCustomer);
-    boolean delete(int id);
-    Customer findById(int id);
+    boolean delete(long id);
+    Customer findById(long id);
     Customer findByName(String name);
     ArrayList<Customer> findAll();
     ArrayList<Customer> findBySearchWords(String searchWords);
